Forum updated

+++ database/migrations/2019_06_05_103015_create_user_teachers_table.php
@@ -42,7 +42,7 @@ class CreateUserTeachersTable extends Migration
             $table->unsignedTinyinteger('status')->length(3);
             $table->rememberToken();
             $table->timestamps();
-            $table->integer('t_authentication')->length(1);
+            $table->tinyInteger('t_authentication')->length(1);
         });
     }
